requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
fastapi==0.103.2
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
pytest==7.4.2
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
